Fix er 3 and 7 buids in latest erbuim (subscribers counter has changed type)

+++ apps/er-coap-07/er-coap-07-observing.c
@@ -167,7 +167,7 @@ coap_remove_observer_by_mid(uip_ipaddr_t *addr, uint16_t port, uint16_t mid)
 }
 /*-----------------------------------------------------------------------------------*/
 void
-coap_notify_observers(resource_t *resource, uint16_t obs_counter, void *notification)
+coap_notify_observers(resource_t *resource, int32_t obs_counter, void *notification)
 {
   coap_packet_t *const coap_res = (coap_packet_t *) notification;
   coap_observer_t* obs = NULL;
